destroy_pod: shut down vm after hyper_term_all
When DESTROYPOD is called by the runtime hyper can call to
hyper_term_all to kill all the process in the VM. After hyper_term_all
is used hyper does not shut down the VM or send and ACK message.
cc-oci-runtime spec to have a message back from hyper.

This patch calls hyper_pod_destroyed after hyper_term_all to sent a
message back to cc-oci-runtime and shutdown the VM.

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>