Do not fail if scanning scsi devices returns failure
In case a file system type is provided for a container which
indicates a device to be used for the container rootfs,
hyperstart scans the scsi devices. Do not fail if the scsi scan
fails.The device to be used may not be a scsi drive and scsi support
may be be built in tke kernel. Do not fail here.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>