exec: Return ERROR if the executed process could not start
The way hyperstart is designed, when we call into EXECCMD, we always
receive an ACK on the ctl channel, even if the process could not start
as expected. Because of this, the runtime does not raise any error
and terminates returning 0.

But this is wrong because it is expected to get the error not only
from the shim, but from the runtime too.

This bug has been found by running CRI-O bats. Specifically with the
test case "ctr execsync failure" from ctr.bats.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>